modelagem_mongodb_veiculos/requirements.txt
pymongo==4.12.1
certifi==2025.4.26
pandas==2.2.3
python-dotenv==1.1.0